% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get the data set used by a gt table}
\usage{
get_data(t)
}
\arguments{
\item{t}{a gt object}
}
\value{
A tibble.
}
\description{
Get the data set used by a gt table
}
\examples{
gt_iris <- gt::gt(iris)
}
