% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_font_names.R
\name{get_font_names}
\alias{get_font_names}
\title{Get the font names used by a gt table}
\usage{
get_font_names(t)
}
\arguments{
\item{t}{a gt object}
}
\value{
A vector.
}
\description{
Get the font names used by a gt table
}
\examples{
gt_iris <- gt::gt(iris)
}
