% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_footnotes.R
\name{get_footnotes}
\alias{get_footnotes}
\title{Get the footnotes used by a gt table}
\usage{
get_footnotes(t)
}
\arguments{
\item{t}{a gt object}
}
\value{
A tibble.
}
\description{
Get the footnotes used by a gt table
}
\examples{
gt_iris <- gt::gt(iris)
}
